+++ platform/mk3/config.mk
@@ -78,6 +78,7 @@ $(HAL_SRC)/wifi_esp8266/wifi_esp8266_device.c \
 $(RCP_SRC)/ADC/ADC.c \
 $(RCP_SRC)/CAN/CAN.c \
 $(RCP_SRC)/CAN/can_mapping.c \
+$(RCP_SRC)/CAN/can_channels.c \
 $(RCP_SRC)/GPIO/GPIO.c \
 $(RCP_SRC)/GPIO/gpioTasks.c \
 $(RCP_SRC)/LED/led.c \
