+++ test/Makefile
@@ -185,6 +185,7 @@ $(MOCK_DIR)/watchdog_device_mock.c \
 $(RCP_SRC)/ADC/ADC.c \
 $(RCP_SRC)/CAN/CAN.c \
 $(RCP_SRC)/CAN/can_mapping.c \
+$(RCP_SRC)/CAN/can_channels.c \
 $(RCP_SRC)/GPIO/GPIO.c \
 $(RCP_SRC)/LED/led.c \
 $(RCP_SRC)/OBD2/OBD2.c \
